MyFirstAIAgent/requirements.txt
agno
groq
openai
python-dotenv
duckduckgo-search
linkedin
pypdf
lancedb
streamlit
pandas
yfinance
